Move SpecificFlowScopeBuilder to dedicated class
Rather than this being logic baked into the `SpecificFlowClient`
init, this is handled by a dedicated class which provides the `user`
scope transformation. It also provides an internal attribute,
`_CLASS_STUB`, for the class-level `scopes` stub object used to
satisfy type checking and provide useful runtime errors.

Effectively, this moves the implementation of these features from
`services.flows` into `scopes.data.flows`.

Because this introduces a dependency from `scopes` onto `_types`, a
typing-time import and string forward references are needed to avoid
circular imports.
