+++ changelog.d/20240813_163828_sirosen_separate_specific_flow_scope_builder.rst
@@ -1,0 +1,8 @@
+Added
+~~~~~
+
+- The scope builder for ``SpecificFlowClient`` is now available for direct
+  access and use via ``globus_sdk.scopes.SpecificFlowScopeBuilder``. Callers can
+  initialize this class with a ``flow_id`` to get a scope builder for a
+  specific flow, e.g., ``SpecificFlowScopeBuilder(flow_id).user``.
+  ``SpecificFlowClient`` now uses this class internally. (:pr:`NUMBER`)
