+++ src/globus_sdk/scopes/data/__init__.py
@@ -1,5 +1,5 @@
 from .auth import AuthScopes
-from .flows import FlowsScopes
+from .flows import FlowsScopes, SpecificFlowScopeBuilder
 from .gcs import GCSCollectionScopeBuilder, GCSEndpointScopeBuilder
 from .groups import GroupsScopes, NexusScopes
 from .search import SearchScopes
@@ -9,6 +9,7 @@ from .transfer import TransferScopes
 __all__ = (
     "AuthScopes",
     "FlowsScopes",
+    "SpecificFlowScopeBuilder",
     "GCSEndpointScopeBuilder",
     "GCSCollectionScopeBuilder",
     "GroupsScopes",
